Rework to use a subclass of NSTableCellView
Instead of trying to insert my own view programmatically, I set up a custom view that corresponds to PlayerView which is a subclass of NSTableCellView. I hooked up two outlets for initiative and name, and then set the values of those outlets in the table view's delegate when it asks for the view.